requirements.txt
torch
torchvision
torchaudio
timm
opencv-python
Pillow
matplotlib
pandas
scikit-learn
streamlit
captum
openslide-python
